requirements.txt
Flask==3.0.3
Pillow==10.4.0
easyocr==1.7.1